♻️ Move end actions to their own file.
